+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 #### Bug Fixes
 - [Predicates / Gates using closures to parameters will cause unexpected behavior.](https://github.com/SAP/chevrotain/issues/221)
+- [Diagrams highlights issues under IE Edge.](https://github.com/SAP/chevrotain/issues/229)
 
 #### Documentation
 - [Example using parametrized sub-rules.](https://github.com/SAP/chevrotain/issues/218)